refactor(ui): extract useHelpTooltip hook into settings/hooks/
